Moved warewulfd commandline program into wwctl
