--- cmd/warewulfd/main.go
@@ -1,12 +1,0 @@
-package main
-
-import (
-	"github.com/hpcng/warewulf/internal/app/warewulfd"
-)
-
-
-func main() {
-	root := warewulfd.GetRootCommand()
-
-	root.Execute()
-}
