+++ internal/app/wwctl/server/start/main.go
@@ -1,8 +1,11 @@
 package start
 
-import "github.com/spf13/cobra"
+import (
+	"github.com/hpcng/warewulf/internal/pkg/warewulfd"
+	"github.com/spf13/cobra"
+)
 
 func CobraRunE(cmd *cobra.Command, args []string) error {
 
-	return nil
+	return warewulfd.RunServer()
 }
