+++ internal/app/wwctl/server/status/main.go
@@ -1,8 +1,12 @@
 package status
 
-import "github.com/spf13/cobra"
+import (
+	"fmt"
+	"github.com/spf13/cobra"
+)
 
 func CobraRunE(cmd *cobra.Command, args []string) error {
+	fmt.Printf("Not implemented yet\n")
 
 	return nil
 }
