+++ internal/app/wwctl/server/stop/main.go
@@ -1,8 +1,11 @@
 package stop
 
-import "github.com/spf13/cobra"
+import (
+	"fmt"
+	"github.com/spf13/cobra"
+)
 
 func CobraRunE(cmd *cobra.Command, args []string) error {
-
+	fmt.Printf("Not implemented yet\n")
 	return nil
 }
