+++ internal/pkg/warewulfd/runtime.go
@@ -1,4 +1,4 @@
-package response
+package warewulfd
 
 import (
 	"fmt"
